safe-ui: inline event_queue into platform as a private submodule
event_queue had no consumers outside of platform — it existed solely
to back the event loop and proxy implementations. Collapsing it into
a private mod event_queue inside platform.rs removes an unnecessary
module boundary without any loss of encapsulation.
